WIP: Removing the need for client
