+++ openghg/cloud/__init__.py
@@ -1,3 +1,3 @@
-__all__ = ["call_function"]
+__all__ = ["call_function", "create_file_package", "create_post_dict"]
 
 from ._call import call_function, create_file_package, create_post_dict
