+++ openghg/retrieve/__init__.py
@@ -1,3 +1,3 @@
+from ._access import get_bc, get_flux, get_footprint, get_obs_surface, get_obs_surface_local
 from ._export import get_ceda_file
-from ._search import search, metadata_lookup
-from ._access import get_obs_surface, get_flux, get_bc, get_footprint
+from ._search import local_search, metadata_lookup, search
