+++ tests/dataobjects/test_searchresults.py
@@ -4,10 +4,9 @@ import numpy as np
 import pandas as pd
 import pytest
 from helpers import get_datapath
-from openghg.client import rank_sources
 from openghg.dataobjects import SearchResults
 from openghg.retrieve import search
-from openghg.store import ObsSurface
+from openghg.store import ObsSurface, rank_sources
 from openghg.util import split_daterange_str
 from pandas import Timestamp
